requirements.txt
fastapi==0.95.0
Jinja2==3.1.2
mysql==0.0.3
mysql-connector==2.2.9
mysqlclient==2.2.0
PyMySQL==1.0.3
python-dotenv==1.0.0
python-multipart==0.0.6
numpy==1.24.4
pandas==1.5.3
requests==2.30.0
schedule==1.1.0
starlette==0.26.1
unicorn==2.0.1.post1
urllib3==1.26.15
uvicorn==0.21.1